Fixed a bug in vjepa student network
