+++ configs/config.yaml
@@ -6,7 +6,7 @@ processed_data_path: './data/processed/'
 # Pytorch
 epochs: 50
 # Game data
-games: ['ms_pacman']
+games: ['breakout']
 
 # Dataset parameters
 batch_size: 32
@@ -26,15 +26,13 @@ use_fixation: True
 maxdist: 25
 sampling_rate: 1000
 
-
-#vjepa 
-stack_size : 4
-Grey_scale_v : True
+#vjepa
+stack_size: 4
+Grey_scale_v: True
 patchx: 21
 patchy: 21
 use_masking: True,
-tubelet_size: 4      # number of frames per tubelet'
-
+tubelet_size: 4 # number of frames per tubelet'
 
 # Data writer parameters
 data_writer:
