+++ src/main.py
@@ -1,6 +1,5 @@
 import matplotlib.pyplot as plt
 import pytorch_lightning as pl
-import torch
 import yaml
 from lightning.pytorch.loggers import TensorBoardLogger
 
@@ -9,7 +8,7 @@ from dataset.pre_process_jepa import ComposePreprocessor, Resize, Stack
 from dataset.torch_dataset import get_torch_dataloaders
 from models.networks import ConvNet, UNet
 from trainers.gaze_predict import GazeTraining
-from trainers.jepa_main import LightningVJEPA
+from trainers.jepa import VJEPA
 from utils import skip_run
 
 # The configuration file
@@ -52,7 +51,7 @@ with skip_run("skip", "gaze_visualization") as check, check():
             ax.set_title(f"Frame {batch_idx},{i} Lable : {label}")
             plt.pause(0.1)
             ax.clear()
-    plt.ioff
+    plt.ioff()
     plt.show()
 
 
@@ -101,131 +100,22 @@ with skip_run("skip", "gaze_prediction_conv_deconv") as check, check():
     )
     trainer.fit(model)
 
-    """
-    # ----------------------------- patch_dim=patch_dim,
-    # Model
-    # -----------------------------
-    model = LightningVJEPA(
-        config=config,
-        student_patch_dim=student_patch_dim,
-        teacher_patch_dim=teacher_patch_dim,
-        embed_dim=768,
-        depth=12,
-        predictor_depth=4,
-        heads=12,
-        mlp_dim=3072,
-        mask_ratio=0.6,
-        lr=1e-4,
-        ema_decay=0.996,
-    )"""
 
-    # -----------------------------
-    # Trainer
-    # -----------------------------
-    trainer = pl.Trainer(
-        accelerator="gpu",
-        devices=1,
-        precision=16,
-        max_epochs=100,
-        logger=logger,
-        enable_progress_bar=True,
-    )
-
-    trainer.fit(model, train_dataloader)
-
-
-with skip_run("skip", "jepa_main") as check, check():
+with skip_run("run", "jepa_training") as check, check():
     game = config["games"][0]
     logger = TensorBoardLogger("tb_logs", name=f"{game}/vjepa_world_model/")
 
     preprocessor = ComposePreprocessor([Resize(config), Stack(config)])
 
     dataloaders = get_torch_dataloaders(game, config, preprocessor=preprocessor)
-
-    train_loader = dataloaders["train"]  # grab the actual DataLoader
-    for x, y in train_loader:
-        print(x.shape)  # gives 32,4,84,84
-
-    train_dataloader = torch.utils.data.DataLoader(
-        jepa_dataset,
-        batch_size=1,
-        shuffle=False,
-        num_workers=1,
-        pin_memory=True,
-    )
-
-    # -----------------------------
-    # Compute patch_dim
-    # -----------------------------
-    T = config["stack_size"]
-    C = 1 if config.get("grey_scale_v", True) else 3
-    a = config["patchx"]
-    b = config["patchy"]
-
-    patch_dim = T * C * a * b
-
-    stack_len = 4
-    C = 1
-    patchx = 21
-    patchy = 21
-
-    student_patch_dim = C * patchx * patchy * (stack_len - 1)  # 1*21*21*3 = 1323
-    teacher_patch_dim = C * patchx * patchy * 1  # 1*21*21*1 = 441
-
-    model = LightningVJEPA(
-        config=config,
-        patch_dim=441,
-        embed_dim=768,
-        depth=12,
-        predictor_depth=4,
-        heads=12,
-        mlp_dim=3072,
-        mask_ratio=0.6,
-        lr=1e-4,
-        ema_decay=0.996,
-    )
-
-
-with skip_run("run", "jepa_main") as check, check():
-    game = config["games"][0]
-    logger = TensorBoardLogger("tb_logs", name=f"{game}/vjepa_world_model/")
-
-    # --------------------------------
-    # Data pipeline (THIS ALREADY WORKS)
-    # --------------------------------
-    preprocessor = ComposePreprocessor(
-        [
-            Resize(config),
-            Stack(config),
-        ]
-    )
-
-    dataloaders = get_torch_dataloaders(
-        game,
-        config,
-        preprocessor=preprocessor,
-    )
-
     train_loader = dataloaders["train"]
 
-    # Sanity check
     for x, y in train_loader:
         print("Train batch shape:", x.shape)  # [32, 4, 84, 84]
         break
 
-    # --------------------------------
-    # Patch configuration (SINGLE source of truth)
-    # --------------------------------
-    C = 1 if config.get("grey_scale_v", True) else 3
-    patchx = config["patchx"]  # 21
-    patchy = config["patchy"]  # 21
-
-    patch_dim = C * patchx * patchy  # 441
-
-    # --------------------------------
-    # Model
-    # --------------------------------
-    model = LightningVJEPA(
+    patch_dim = 1 if config.get("grey_scale_v", True) else 3
+    model = VJEPA(
         config=config,
         patch_dim=patch_dim,
         embed_dim=768,
@@ -238,14 +128,12 @@ with skip_run("run", "jepa_main") as check, check():
         ema_decay=0.996,
     )
 
-    # --------------------------------
-    # Trainer
-    # --------------------------------
     trainer = pl.Trainer(
         logger=logger,
         max_epochs=config["epochs"],
         accelerator="auto",
         devices="auto",
+        precision="bf16-mixed",
         log_every_n_steps=10,
     )
